update: restructure files correctly
