--- packages/shared/babel.config.js
@@ -1,16 +1,0 @@
-module.exports = {
-  presets: [
-    ["@babel/preset-env", { targets: { node: "current" } }],
-    "@babel/preset-react",
-  ],
-  plugins: [
-    [
-      "module-resolver",
-      {
-        alias: {
-          tailwindcss: "../shared/node_modules/tailwindcss",
-        },
-      },
-    ],
-  ],
-};
