--- packages/shared/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "theme",
-  "version": "1.0.0",
-  "main": "index.js",
-  "license": "MIT",
-  "dependencies": {
-    "@heroicons/react": "^2.0.11",
-    "autoprefixer": "^10.4.14",
-    "flowbite": "^1.5.3",
-    "flowbite-react": "^0.1.11",
-    "framer-motion": "^7.5.1",
-    "postcss": "^8.4.21",
-    "tailwindcss": "^3.3.1"
-  },
-  "devDependencies": {
-    "@babel/preset-env": "^7.21.4",
-    "@babel/preset-react": "^7.18.6"
-  }
-}
